canbus: fix compile warning. (#2425)

+++ modules/canbus/vehicle/gem/protocol/turn_cmd_63.cc
@@ -57,7 +57,7 @@ Turncmd63* Turncmd63::set_turn_signal_cmd(
 // 'bit': 7, 'type': 'enum', 'order': 'motorola', 'physical_unit': ''}
 void Turncmd63::set_p_turn_signal_cmd(
     uint8_t* data, Turn_cmd_63::Turn_signal_cmdType turn_signal_cmd) {
-  int x = turn_signal_cmd;
+  uint8_t x = turn_signal_cmd;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 0, 8);
